Check command miyagi compatibility
